pyproject.toml
[project]
name = "fastapi-keycloak-auth"
version = "1.0.3"
description = ""
authors = [
    { name = "paulkczw", email = "paul+development@kczw.de" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.128.1,<0.129.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
]

[dependency-groups]
dev = [
    "uvicorn (>=0.40.0,<0.41.0)"
]
test = [
    { include-group = "dev" },
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=0.25.0,<1.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "pytest-mock (>=3.15.1,<4.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{ include = "fastapi_keycloak_auth", from = "src" }]

[tool.poetry.group.test.dependencies]
pytest-asyncio = "^1.3.0"
cryptography = "^46.0.5"

